feature: divide evaluation.py into agent and eval
